Creating update_mode.py for first callback
