+++ src/app.py
@@ -16,6 +16,7 @@ parent_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(parent_dir)
 
 from src.data.preprocessing import load, widget_inputs
+from src.callbacks.update_mode import update_mode_callback
 
 ### --- LOAD AND PREPROCESS DATA ---
 
@@ -27,21 +28,7 @@ alt.data_transformers.enable("vegafusion")
 app = dash.Dash(__name__, external_stylesheets=[dbc.themes.CERULEAN], title="Commuting Insights Dashboard")
 server = app.server
 
-@app.callback(
-    Output("mode-dropdown", "options"),
-    [Input("cd-dropdown", "value")]
-)
-def update_mode_options(selected_cd):
-    if not selected_cd:
-        return dropdown_options
-    # Filter for the selected CD and nonzero average commute time
-    df_cd = df[(df["GEO"] == selected_cd) & (df["AverageCommuteTime"] > 0)]
-    # Get the unique modes that appear in the CD
-    modes = df_cd["Main mode of commuting (21)"].unique()
-    # Only keep those modes that are in the allowed available_modes list
-    valid_modes = [m for m in modes if m in available_modes]
-    options = [{"label": m, "value": m} for m in valid_modes]
-    return options
+update_mode_callback(df, available_modes, dropdown_options)
 
 @app.callback(
     Output("cd-dropdown", "options"),
